fix: rename and ui setup in drawers
