+++ apps/directorate-of-equality-web/src/components/reports/tabs/TabContent.tsx
